@@ -26,7 +26,7 @@ import com.google.maps.model.TransitMode;
 import com.google.maps.model.TransitRoutingPreference;
 import com.google.maps.model.TravelMode;
 import com.google.maps.model.Unit;
-import org.joda.time.ReadableInstant;
+import java.time.Instant;
 
 /** A request to the Distance Matrix API. */
 public class DistanceMatrixApiRequest
@@ -44,8 +44,7 @@ public class DistanceMatrixApiRequest
     if (!params().containsKey("destinations")) {
       throw new IllegalArgumentException("Request must contain 'destinations'");
     }
-    if (TravelMode.TRANSIT.toString().equals(params().get("mode"))
-        && (params().containsKey("arrival_time") && params().containsKey("departure_time"))) {
+    if (params().containsKey("arrival_time") && params().containsKey("departure_time")) {
       throw new IllegalArgumentException(
           "Transit request must not contain both a departureTime and an arrivalTime");
     }
@@ -157,8 +156,8 @@ public class DistanceMatrixApiRequest
    * @param departureTime The time of departure.
    * @return Returns this {@code DistanceMatrixApiRequest} for call chaining.
    */
-  public DistanceMatrixApiRequest departureTime(ReadableInstant departureTime) {
-    return param("departure_time", Long.toString(departureTime.getMillis() / 1000L));
+  public DistanceMatrixApiRequest departureTime(Instant departureTime) {
+    return param("departure_time", Long.toString(departureTime.toEpochMilli() / 1000L));
   }
 
   /**
@@ -179,8 +178,8 @@ public class DistanceMatrixApiRequest
    * @param arrivalTime The preferred arrival time.
    * @return Returns this {@code DistanceMatrixApiRequest} for call chaining.
    */
-  public DistanceMatrixApiRequest arrivalTime(ReadableInstant arrivalTime) {
-    return param("arrival_time", Long.toString(arrivalTime.getMillis() / 1000L));
+  public DistanceMatrixApiRequest arrivalTime(Instant arrivalTime) {
+    return param("arrival_time", Long.toString(arrivalTime.toEpochMilli() / 1000L));
   }
 
   /**
