@@ -63,7 +63,7 @@ public class OkHttpRequestHandler implements GeoApiContext.RequestHandler {
     Request req =
         new Request.Builder().get().header("User-Agent", userAgent).url(hostName + url).build();
 
-    return new OkHttpPendingResult<T, R>(
+    return new OkHttpPendingResult<>(
         req, client, clazz, fieldNamingPolicy, errorTimeout, maxRetries, exceptionsAllowedToRetry);
   }
 
@@ -86,12 +86,14 @@ public class OkHttpRequestHandler implements GeoApiContext.RequestHandler {
             .url(hostName + url)
             .build();
 
-    return new OkHttpPendingResult<T, R>(
+    return new OkHttpPendingResult<>(
         req, client, clazz, fieldNamingPolicy, errorTimeout, maxRetries, exceptionsAllowedToRetry);
   }
 
+  @Override
   public void shutdown() {
     executorService.shutdown();
+    client.connectionPool().evictAll();
   }
 
   /** Builder strategy for constructing an {@code OkHTTPRequestHandler}. */
@@ -108,34 +110,39 @@ public class OkHttpRequestHandler implements GeoApiContext.RequestHandler {
     }
 
     @Override
-    public void connectTimeout(long timeout, TimeUnit unit) {
+    public Builder connectTimeout(long timeout, TimeUnit unit) {
       builder.connectTimeout(timeout, unit);
+      return this;
     }
 
     @Override
-    public void readTimeout(long timeout, TimeUnit unit) {
+    public Builder readTimeout(long timeout, TimeUnit unit) {
       builder.readTimeout(timeout, unit);
+      return this;
     }
 
     @Override
-    public void writeTimeout(long timeout, TimeUnit unit) {
+    public Builder writeTimeout(long timeout, TimeUnit unit) {
       builder.writeTimeout(timeout, unit);
+      return this;
     }
 
     @Override
-    public void queriesPerSecond(int maxQps) {
+    public Builder queriesPerSecond(int maxQps) {
       dispatcher.setMaxRequests(maxQps);
       dispatcher.setMaxRequestsPerHost(maxQps);
       rateLimitExecutorService.setQueriesPerSecond(maxQps);
+      return this;
     }
 
     @Override
-    public void proxy(Proxy proxy) {
+    public Builder proxy(Proxy proxy) {
       builder.proxy(proxy);
+      return this;
     }
 
     @Override
-    public void proxyAuthentication(String proxyUserName, String proxyUserPassword) {
+    public Builder proxyAuthentication(String proxyUserName, String proxyUserPassword) {
       final String userName = proxyUserName;
       final String password = proxyUserPassword;
 
@@ -151,6 +158,19 @@ public class OkHttpRequestHandler implements GeoApiContext.RequestHandler {
                   .build();
             }
           });
+      return this;
+    }
+
+    /**
+     * Gets a reference to the OkHttpClient.Builder used to build the OkHttpRequestHandler's
+     * internal OkHttpClient. This allows you to fully customize the OkHttpClient that the resulting
+     * OkHttpRequestHandler will make HTTP requests through.
+     *
+     * @return OkHttpClient.Builder that will produce the OkHttpClient used by the
+     *     OkHttpRequestHandler built by this.
+     */
+    public OkHttpClient.Builder okHttpClientBuilder() {
+      return builder;
     }
 
     @Override
