@@ -42,13 +42,13 @@ import com.google.maps.model.PriceLevel;
 import com.google.maps.model.TravelMode;
 import java.io.IOException;
 import java.nio.charset.Charset;
+import java.time.Instant;
+import java.time.LocalTime;
+import java.time.ZonedDateTime;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
-import org.joda.time.DateTime;
-import org.joda.time.Instant;
-import org.joda.time.LocalTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -166,22 +166,17 @@ public class GaePendingResult<T, R extends ApiResponse<T>> implements PendingRes
 
     Gson gson =
         new GsonBuilder()
-            .registerTypeAdapter(DateTime.class, new DateTimeAdapter())
+            .registerTypeAdapter(ZonedDateTime.class, new ZonedDateTimeAdapter())
             .registerTypeAdapter(Distance.class, new DistanceAdapter())
             .registerTypeAdapter(Duration.class, new DurationAdapter())
             .registerTypeAdapter(Fare.class, new FareAdapter())
             .registerTypeAdapter(LatLng.class, new LatLngAdapter())
             .registerTypeAdapter(
-                AddressComponentType.class,
-                new SafeEnumAdapter<AddressComponentType>(AddressComponentType.UNKNOWN))
-            .registerTypeAdapter(
-                AddressType.class, new SafeEnumAdapter<AddressType>(AddressType.UNKNOWN))
-            .registerTypeAdapter(
-                TravelMode.class, new SafeEnumAdapter<TravelMode>(TravelMode.UNKNOWN))
-            .registerTypeAdapter(
-                LocationType.class, new SafeEnumAdapter<LocationType>(LocationType.UNKNOWN))
-            .registerTypeAdapter(
-                RatingType.class, new SafeEnumAdapter<RatingType>(RatingType.UNKNOWN))
+                AddressComponentType.class, new SafeEnumAdapter<>(AddressComponentType.UNKNOWN))
+            .registerTypeAdapter(AddressType.class, new SafeEnumAdapter<>(AddressType.UNKNOWN))
+            .registerTypeAdapter(TravelMode.class, new SafeEnumAdapter<>(TravelMode.UNKNOWN))
+            .registerTypeAdapter(LocationType.class, new SafeEnumAdapter<>(LocationType.UNKNOWN))
+            .registerTypeAdapter(RatingType.class, new SafeEnumAdapter<>(RatingType.UNKNOWN))
             .registerTypeAdapter(DayOfWeek.class, new DayOfWeekAdapter())
             .registerTypeAdapter(PriceLevel.class, new PriceLevelAdapter())
             .registerTypeAdapter(Instant.class, new InstantAdapter())
