@@ -79,6 +79,20 @@ public class CellTower implements Serializable {
   /** The timing advance value. */
   public Integer timingAdvance = null;
 
+  @Override
+  public String toString() {
+    return String.format(
+        "[CellTower: cellId=%s, locationAreaCode=%s, mobileCountryCode=%s, "
+            + "mobileNetworkCode=%s, age=%s, signalStrength=%s, timingAdvance=%s]",
+        cellId,
+        locationAreaCode,
+        mobileCountryCode,
+        mobileNetworkCode,
+        age,
+        signalStrength,
+        timingAdvance);
+  }
+
   public static class CellTowerBuilder {
     private Integer _cellId = null;
     private Integer _locationAreaCode = null;
@@ -101,37 +115,37 @@ public class CellTower implements Serializable {
     }
 
     public CellTowerBuilder CellId(int newCellId) {
-      this._cellId = new Integer(newCellId);
+      this._cellId = newCellId;
       return this;
     }
 
     public CellTowerBuilder LocationAreaCode(int newLocationAreaCode) {
-      this._locationAreaCode = new Integer(newLocationAreaCode);
+      this._locationAreaCode = newLocationAreaCode;
       return this;
     }
 
     public CellTowerBuilder MobileCountryCode(int newMobileCountryCode) {
-      this._mobileCountryCode = new Integer(newMobileCountryCode);
+      this._mobileCountryCode = newMobileCountryCode;
       return this;
     }
 
     public CellTowerBuilder MobileNetworkCode(int newMobileNetworkCode) {
-      this._mobileNetworkCode = new Integer(newMobileNetworkCode);
+      this._mobileNetworkCode = newMobileNetworkCode;
       return this;
     }
 
     public CellTowerBuilder Age(int newAge) {
-      this._age = new Integer(newAge);
+      this._age = newAge;
       return this;
     }
 
     public CellTowerBuilder SignalStrength(int newSignalStrength) {
-      this._signalStrength = new Integer(newSignalStrength);
+      this._signalStrength = newSignalStrength;
       return this;
     }
 
     public CellTowerBuilder TimingAdvance(int newTimingAdvance) {
-      this._timingAdvance = new Integer(newTimingAdvance);
+      this._timingAdvance = newTimingAdvance;
       return this;
     }
   }
