@@ -15,8 +15,11 @@
 
 package com.google.maps.model;
 
+import static com.google.maps.internal.StringJoin.join;
+
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -76,6 +79,34 @@ public class GeolocationPayload implements Serializable {
   /** An array of WiFi access point objects. See {@link com.google.maps.model.WifiAccessPoint}. */
   public WifiAccessPoint[] wifiAccessPoints;
 
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder("[GeolocationPayload");
+    List<String> elements = new ArrayList<>();
+    if (homeMobileCountryCode != null) {
+      elements.add("homeMobileCountryCode=" + homeMobileCountryCode);
+    }
+    if (homeMobileNetworkCode != null) {
+      elements.add("homeMobileNetworkCode=" + homeMobileNetworkCode);
+    }
+    if (radioType != null) {
+      elements.add("radioType=" + radioType);
+    }
+    if (carrier != null) {
+      elements.add("carrier=" + carrier);
+    }
+    elements.add("considerIp=" + considerIp);
+    if (cellTowers != null && cellTowers.length > 0) {
+      elements.add("cellTowers=" + Arrays.toString(cellTowers));
+    }
+    if (wifiAccessPoints != null && wifiAccessPoints.length > 0) {
+      elements.add("wifiAccessPoints=" + Arrays.toString(wifiAccessPoints));
+    }
+    sb.append(join(", ", elements));
+    sb.append("]");
+    return sb.toString();
+  }
+
   public static class GeolocationPayloadBuilder {
     private Integer _homeMobileCountryCode = null;
     private Integer _homeMobileNetworkCode = null;
@@ -83,9 +114,9 @@ public class GeolocationPayload implements Serializable {
     private String _carrier = null;
     private Boolean _considerIp = null;
     private CellTower[] _cellTowers = null;
-    private List<CellTower> _addedCellTowers = new ArrayList<CellTower>();
+    private List<CellTower> _addedCellTowers = new ArrayList<>();
     private WifiAccessPoint[] _wifiAccessPoints = null;
-    private List<WifiAccessPoint> _addedWifiAccessPoints = new ArrayList<WifiAccessPoint>();
+    private List<WifiAccessPoint> _addedWifiAccessPoints = new ArrayList<>();
 
     public GeolocationPayload createGeolocationPayload() {
       // if wifi access points have been added individually...
@@ -110,12 +141,12 @@ public class GeolocationPayload implements Serializable {
     }
 
     public GeolocationPayloadBuilder HomeMobileCountryCode(int newHomeMobileCountryCode) {
-      this._homeMobileCountryCode = new Integer(newHomeMobileCountryCode);
+      this._homeMobileCountryCode = newHomeMobileCountryCode;
       return this;
     }
 
     public GeolocationPayloadBuilder HomeMobileNetworkCode(int newHomeMobileNetworkCode) {
-      this._homeMobileNetworkCode = new Integer(newHomeMobileNetworkCode);
+      this._homeMobileNetworkCode = newHomeMobileNetworkCode;
       return this;
     }
 
@@ -130,7 +161,7 @@ public class GeolocationPayload implements Serializable {
     }
 
     public GeolocationPayloadBuilder ConsiderIp(boolean newConsiderIp) {
-      this._considerIp = new Boolean(newConsiderIp);
+      this._considerIp = newConsiderIp;
       return this;
     }
 
