@@ -25,6 +25,7 @@ import com.google.maps.model.LatLng;
 import com.google.maps.model.SnappedPoint;
 import com.google.maps.model.SnappedSpeedLimitResponse;
 import com.google.maps.model.SpeedLimit;
+import java.util.Arrays;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
@@ -62,10 +63,11 @@ public class RoadsApiIntegrationTest {
           };
       SnappedPoint[] points = RoadsApi.snapToRoads(sc.context, false, path).await();
 
+      assertNotNull(Arrays.toString(points));
       sc.assertParamValue(join('|', path), "path");
       sc.assertParamValue("false", "interpolate");
       assertEquals(7, points.length);
-      assertEquals(-33.865233402568428, points[0].location.lat, 0.0001);
+      assertEquals(-33.86523340256843, points[0].location.lat, 0.0001);
       assertEquals(151.19288612197704, points[0].location.lng, 0.0001);
       assertEquals("ChIJjXkMCDauEmsRp5xab4Ske6k", points[0].placeId);
     }
@@ -86,6 +88,7 @@ public class RoadsApiIntegrationTest {
           };
       SpeedLimit[] speeds = RoadsApi.speedLimits(sc.context, path).await();
 
+      assertNotNull(Arrays.toString(speeds));
       assertEquals("/v1/speedLimits", sc.path());
       sc.assertParamValue(join('|', path), "path");
       assertEquals(7, speeds.length);
@@ -112,6 +115,7 @@ public class RoadsApiIntegrationTest {
           };
       SpeedLimit[] speeds = RoadsApi.speedLimits(sc.context, path).await();
 
+      assertNotNull(Arrays.toString(speeds));
       assertEquals("/v1/speedLimits", sc.path());
       sc.assertParamValue(join('|', path), "path");
       assertEquals(7, speeds.length);
@@ -134,6 +138,7 @@ public class RoadsApiIntegrationTest {
           };
       SpeedLimit[] speeds = RoadsApi.speedLimits(sc.context, placeIds).await();
 
+      assertNotNull(Arrays.toString(speeds));
       assertEquals("/v1/speedLimits", sc.path());
       assertEquals(3, speeds.length);
       assertEquals("ChIJc0BrC2EE9YgR71DvaFzNgrA", speeds[2].placeId);
@@ -159,6 +164,7 @@ public class RoadsApiIntegrationTest {
           };
       SnappedSpeedLimitResponse response = RoadsApi.snappedSpeedLimits(sc.context, path).await();
 
+      assertNotNull(response.toString());
       assertEquals("/v1/speedLimits", sc.path());
       sc.assertParamValue(join('|', path), "path");
       assertEquals(path.length, response.snappedPoints.length);
@@ -181,10 +187,11 @@ public class RoadsApiIntegrationTest {
           };
       SnappedPoint[] points = RoadsApi.nearestRoads(sc.context, path).await();
 
+      assertNotNull(Arrays.toString(points));
       assertEquals("/v1/nearestRoads", sc.path());
       sc.assertParamValue(join('|', path), "points");
       assertEquals(13, points.length);
-      assertEquals(-33.865436156120467, points[0].location.lat, 0.0001);
+      assertEquals(-33.86543615612047, points[0].location.lat, 0.0001);
       assertEquals(151.1930101572747, points[0].location.lng, 0.0001);
       assertEquals("ChIJ0XXACjauEmsRUduC5Wd9ARM", points[0].placeId);
     }
